Struct sketchy::BloomFilter

pub struct BloomFilter<E> {
    // some fields omitted
}

A Bloom filter is a space-efficient probabilistic data structure that is used to test whether an element is a member of a set. False positive matches are possible, but false negatives are not.

use sketchy::BloomFilter;

// Create a filter which can handle 100K elements with a 1% maximum
// probability of a false positive.
let mut filter = BloomFilter::new(100_000, 0.01);

filter.insert("one");
filter.insert("two");

assert!(filter.contains(&"one"));

Methods

impl<E: Hash> BloomFilter<E>

fn new(n: usize, max_false_pos_prob: f64) -> BloomFilter<E>

Creates a new BloomFilter instance, tuned for a population of n elements with the given upper bound of the probability of false positives.

fn insert(&mut self, e: E)

Adds a value to the set.

fn contains(&mut self, e: &E) -> bool

Returns true if the set probably contains the given element.

fn merge(&mut self, other: &BloomFilter<E>) -> bool

Merges the contents of the given BloomFilter into self. Both filters must have the same parameters. Returns true if self changed.

Panics

Panics if the bloom filters have different parameters.