Struct sketchy::CountMinSketch

pub struct CountMinSketch<E> {
    // some fields omitted
}

A Count-Min Sketch is a probabilistic data structure which provides estimates of the frequency of elements in a data stream. It is parameterized with the type of elements.

use sketchy::CountMinSketch;

let mut cms = CountMinSketch::with_confidence(0.001, 0.99);
cms.insert("one hundred");
cms.insert_n("one hundred", 100);

assert_eq!(cms.estimate(&"one hundred"), 101);

Methods

impl<E: Hash> CountMinSketch<E>

fn with_confidence(epsilon: f64, confidence: f64) -> CountMinSketch<E>

Returns a CountMinSketch which provides frequency estimates where the error is within a factor of epsilon with the given confidence.

fn new(depth: usize, width: usize) -> CountMinSketch<E>

Returns a CountMinSketch with the given depth and width.

fn insert(&mut self, e: E)

Adds a value to the sketch.

fn insert_n(&mut self, e: E, n: u64)

Adds multiple instances of a value to the sketch.

fn estimate(&self, e: &E) -> u64

Estimates the frequency of the given element.

fn estimate_mean(&self, e: E, n: u64) -> u64

Estimates the frequency of the given element using the Count-Mean-Min algorithm, which performs better on data sets which aren't highly skewed.

fn merge(&mut self, v: &CountMinSketch<E>)

Merges another CountMinSketch into self.