Struct sketchy::ReservoirSample

pub struct ReservoirSample<E> {
    // some fields omitted
}

A reservoir sample maintains a sample of K elements, selected uniformly and at random from a stream. This implementation uses Vitter's Algorithm R with thread-local PRNGs.

use sketchy::ReservoirSample;

let mut res = ReservoirSample::new(2);

for v in vec!["one", "two", "three"] {
    res.insert(v);
}

assert_eq!(res.elements().len(), 2);

Methods

impl<E: Clone> ReservoirSample<E>

fn new(size: usize) -> ReservoirSample<E>

Returns a new ReservoirSample of the given size.

fn insert(&mut self, e: E)

Inserts the given element into the sample.

fn elements(self) -> Vec<E>

Returns the recorded elements in the sample.