Struct sketchy::TopK

pub struct TopK<E> {
    // some fields omitted
}

A Top-K heap is a probabilistic data structure which uses a Count-Min Sketch to calculate the top K elements in a data stream with the highest frequency.

use sketchy::{CountMinSketch, TopK};

let cms = CountMinSketch::with_confidence(0.0001, 0.99);
let mut topk = TopK::new(5, 0.05, cms);

for i in 1..10000 {
    topk.insert(i % 1000); // an uncommon item
    topk.insert(-100); // a common item
}

assert_eq!(topk.elements(), vec![-100]);

Methods

impl<E: Eq + Hash + Copy> TopK<E>

fn new(k: usize, min: f64, cms: CountMinSketch<E>) -> TopK<E>

Returns a TopK which will track k elements with at least min frequency ((0,1)) using the given CountMinSketch.

fn insert(&mut self, e: E)

Adds a value to the heap.

fn elements(&mut self) -> Vec<E>

Returns a vector of the top K elements, in reverse order of frequency.

fn shrink_to_fit(&mut self)

Shrinks the heap as much as possible while still retaining the top K elements. Should be called periodically to filter out false positives as the data stream changes.